Move components out of pages
